Moved common startup code to new module
